#12  fix chrome login in demo mode

+++ app/models/file/ServerFile.java
@@ -72,6 +72,10 @@ public class ServerFile extends Model {
 
     public static void deleteFile(ServerFile file) {
         File fileDir = new File(file.directoryPath);
+        if (!fileDir.exists()) {
+            Ebean.delete(file);
+            return;
+        }
         File[] files = fileDir.listFiles();
         if (files == null) {
             fileDir.delete();
